Standardized export syntax.
The existing pattern for this repo is to declare an export where the value is defined, this ensures that reading the value immediately tells you if its public API for that given module, or if it is an internal detail.

Someday we will move to mjs, and the syntax becomes a little nicer.