v0.6.0 - Round end of support to end of quarter
Rather than setting end of support dates for dependency packages to be exactly N quarters from the previous release, allow the entire rest of the last quarter to upgrade. For example, if a patch version is released on May 5th, the end of life date for the previous patch version should be September 30th, not August 5th.